requirements.txt
packaging==24.2
pillow==11.2.1
pytesseract==0.3.13
